Detached telajax from crate telamon
