+++ Cargo.toml
@@ -96,6 +96,7 @@ flate2 = "1.0.2"
 dot = "0.1"
 csv = "1"
 utils = {package = "telamon-utils", path = "telamon-utils"}
+telajax-bind = {package = "telajax-bind", path = "telajax-bind"}
 
 [dependencies.tokio-timer]
 git = "https://github.com/ulysseB/tokio-timer.git"
